fcp-microservices/services/workbench-services/account-filter/requirements.txt
snowflake-sqlalchemy
snowflake-connector-python
sqlalchemy
boto3
python-dotenv
pandas